Modularize json for vertx
