+++ sqrl-execute/sqrl-execute-http/sqrl-execute-http-vertx/src/main/java/com/datasqrl/graphql/GraphQLServer.java
@@ -16,6 +16,7 @@ import com.datasqrl.graphql.server.BuildGraphQLEngine;
 import com.datasqrl.graphql.server.Model.MutationCoords;
 import com.datasqrl.graphql.server.Model.RootGraphqlModel;
 import com.datasqrl.graphql.server.Model.SubscriptionCoords;
+import com.datasqrl.graphql.type.JsonTypeFactory;
 import com.datasqrl.io.impl.jdbc.JdbcDataSystemConnector;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import graphql.GraphQL;
@@ -42,6 +43,7 @@ import io.vertx.sqlclient.PoolOptions;
 import io.vertx.sqlclient.SqlClient;
 import java.io.File;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -175,7 +177,7 @@ public class GraphQLServer extends AbstractVerticle {
   public GraphQL createGraphQL(SqlClient client, Promise<Void> startPromise) {
     try {
       GraphQL graphQL = model.accept(
-          new BuildGraphQLEngine(),
+          new BuildGraphQLEngine(List.of(new JsonTypeFactory())),
           new VertxContext(new VertxJdbcClient(client), constructSinkProducers(model, vertx),
               constructSubscriptions(model, vertx), canonicalizer));
       return graphQL;
